Result Obtained for analysis
This result contains three column namely 'Keywords' which list out all present keywords in pdf file, 'Normalized Weightage' which is used to know the importance of keywords in that document. This importance is calculated by counting number of occurrence of all keywords in that document. Thus weightage is calculated from only one document and 'No of occurrence' is the number of occurrence of keywords in that document.